port voice engine

+++ voice_engine/pyaudio_source.py
@@ -29,7 +29,7 @@ class Source(Element):
         self.pyaudio_instance = pyaudio.PyAudio()
 
         formats = [pyaudio.paInt8, pyaudio.paInt16, pyaudio.paInt24, pyaudio.paInt32]
-        width = formats[bits_per_sample / 8 - 1]
+        width = formats[int(bits_per_sample / 8 - 1)]
 
         # Search device by name
         if device_name:
